Revert "chore: Corrigido erro no cadastro de instituições na assinatura"
This reverts commit 906c5b294581ca4fb816e746e287ffeaa0e09a89.

+++ backend/proofchain/src/main/java/com/proofchain/Dtos/request/NewInstituitionRequestDto.java
@@ -24,5 +24,5 @@ public class NewInstituitionRequestDto {
     private String email;
     @Size(min=8)
     private String password;
-    private Long idPlan;
+    private int idPlan;
 }
